Separate drawing and OpenVR into their own processes
This change separates the OpenVRHelper and DrawHelper into their own processes. This requires additional synchronouse bootstrapping to share resources between the two, as documented further in SampleVRO.cpp.
The core of this change is the call to IVROverlay::SetOverlayRenderingPid, which enables drawing to the Overlay texture in a separate process.